# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Creates a Supabase user and the corresponding application User row. Rate-limited to 10/hour/IP to blunt enumeration and spam signup loops.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session tokens for the new user."
          },
          "400": {
            "description": "Validation error."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Register a new user with email + password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Returns Supabase access/refresh tokens. Rate-limited 10/min/IP.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session."
          },
          "401": {
            "description": "Invalid credentials."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Email + password login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/google": {
      "post": {
        "description": "Exchanges a Google ID token for a Supabase session. Optional `invite_code` attaches the new user to a coach in the same call.",
        "operationId": "AuthController_googleAuth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated session."
          },
          "401": {
            "description": "Invalid Google token."
          }
        },
        "summary": "Google OAuth exchange",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/attach-invite-code": {
      "post": {
        "description": "Idempotent — re-running with the same code is a no-op.",
        "operationId": "AuthController_attachInviteCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachInviteCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attached to coach."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "description": "Invite code not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Attach the caller to a coach via invite code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/select-role": {
      "post": {
        "description": "Only `student` is honored — coach elevation must be done by an OWNER admin. A `coach` value will be rejected with 403.",
        "operationId": "AuthController_selectRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role applied."
          },
          "403": {
            "description": "Role elevation rejected."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Select role + optionally attach invite code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/validate-invite-code": {
      "post": {
        "description": "Returns `{ valid, coach_id?, coach_name? }`. Format-validates the code before any DB lookup so the response is identical for malformed input. Rate-limited 20/min/IP.",
        "operationId": "AuthController_validateInviteCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateInviteCodePublicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite code preview result."
          },
          "400": {
            "description": "Invite code format is invalid."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Preview an invite code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "description": "Always returns 200 to avoid leaking whether an email is registered.",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email dispatched if user exists."
          }
        },
        "summary": "Trigger a password-reset email",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Caller profile."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/become-coach": {
      "post": {
        "description": "Requires re-entering the current password. Gated by COACH_SELF_ELEVATION_ENABLED.",
        "operationId": "AuthController_becomeCoach",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BecomeCoachDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coach role granted."
          },
          "401": {
            "description": "Wrong password."
          },
          "403": {
            "description": "Self-elevation disabled."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Self-elevate to coach role",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signup-with-code": {
      "post": {
        "description": "Behind COACH_CODE_GATE_ENABLED=true the invite_code is required.",
        "operationId": "AuthController_signupWithCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupWithCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New session, attached to coach."
          },
          "400": {
            "description": "Invite code missing or invalid."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Signup including a coach invite code in one call",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Growth Project API",
    "description": "HTTP API for The Growth Project — coach/client SaaS platform covering auth, billing, messaging, check-ins, meal plans, workouts, habits, and admin/federation surfaces. Consumed by the React Native mobile app, the coach console BFF, and partner integrations.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.trygrowthproject.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase-issued JWT. Send as `Authorization: Bearer <token>`. Tokens are verified against the Supabase JWKS in JwtAuthGuard."
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "correct horse battery staple"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "phone": {
            "type": "string",
            "example": "+15555550100"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "correct horse battery staple"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "GoogleAuthDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Google ID token from the mobile or web SDK.",
            "minLength": 10
          },
          "invite_code": {
            "type": "string",
            "description": "Optional coach invite code attached on the same call.",
            "example": "GP-A1B2C3",
            "maxLength": 32
          }
        },
        "required": [
          "token"
        ]
      },
      "AttachInviteCodeDto": {
        "type": "object",
        "properties": {
          "invite_code": {
            "type": "string",
            "example": "GP-A1B2C3",
            "minLength": 3,
            "maxLength": 32
          }
        },
        "required": [
          "invite_code"
        ]
      },
      "SelectRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "coach",
              "student"
            ],
            "example": "student"
          },
          "coach_code": {
            "type": "string",
            "deprecated": true,
            "description": "Legacy alias of invite_code; prefer invite_code on new clients.",
            "maxLength": 32
          },
          "invite_code": {
            "type": "string",
            "example": "GP-A1B2C3",
            "maxLength": 32
          }
        },
        "required": [
          "role"
        ]
      },
      "ValidateInviteCodePublicDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Invite code to preview. Format-validated server-side.",
            "example": "GP-A1B2C3",
            "maxLength": 256
          }
        },
        "required": [
          "code"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "BecomeCoachDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Current password — required to prevent accidental escalation.",
            "minLength": 8
          }
        },
        "required": [
          "password"
        ]
      },
      "SignupWithCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "phone": {
            "type": "string",
            "example": "+15555550100"
          },
          "invite_code": {
            "type": "string",
            "description": "Required when COACH_CODE_GATE_ENABLED=true; otherwise optional.",
            "example": "GP-A1B2C3",
            "maxLength": 32
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
...
    }
  }
}